class per file and some more tidy up
