+++ app/uk/gov/hmrc/eventhub/subscription/stream/SubscriberEventSource.scala
@@ -14,22 +14,12 @@
  * limitations under the License.
  */
 
-package uk.gov.hmrc.eventhub.subscriptions.http
+package uk.gov.hmrc.eventhub.subscription.stream
 
-import akka.http.scaladsl.model.{ContentTypes, HttpEntity, HttpRequest}
-import play.api.libs.json.Json
-import uk.gov.hmrc.eventhub.model.{Event, Subscriber}
+import akka.NotUsed
+import akka.stream.scaladsl.Source
+import uk.gov.hmrc.eventhub.model.Event
 
-trait HttpEventRequestBuilder {
-  def build(subscriber: Subscriber, event: Event): HttpRequest =
-    HttpRequest.apply(
-      method = subscriber.httpMethod,
-      uri = subscriber.uri,
-      entity = HttpEntity(
-        ContentTypes.`application/json`,
-        Json.toJson(event).toString()
-      )
-    )
+trait SubscriberEventSource {
+  def source: Source[Event, NotUsed]
 }
-
-object HttpEventRequestBuilder extends HttpEventRequestBuilder
